+++ src/server/server.js
@@ -5,7 +5,7 @@ const cors = require('cors');
 
 const app = express();
 const PORT = 8080;
-const OldScaffolderRouter = require("./oldScaffolder.js");
+const OldScaffolderRouter = require("./routers/oldScaffolder.js");
 
 require('dotenv').config();
